+++ examples/compare_ISMN_ASCAT.py
@@ -37,7 +37,7 @@ Created on Aug 8, 2013
 @author: Christoph Paulik Christoph.Paulik@geo.tuwien.ac.at
 '''
 
-import pytesmo.io.ismn.interface as ismn
+import ismn.interface as ismn
 import ascat
 import pytesmo.temporal_matching as temp_match
 import pytesmo.scaling as scaling